requirements.txt
Flask
Pillow
opencv-python-headless
numpy